# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Zenith Backend API",
    "description": "Paper-trading API for the Zenith options protocol.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/chain": {
      "get": {
        "tags": [
          "crate"
        ],
        "operationId": "get_chain",
        "parameters": [
          {
            "name": "underlying",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "expiry_days",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OptionChainEntry"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Unknown underlying"
          }
        }
      }
    },
...
    "/api/v1/iv": {
      "get": {
        "tags": [
          "crate"
        ],
        "operationId": "get_implied_vol",
        "parameters": [
          {
            "name": "underlying",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "strike",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "expiry_days",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "option_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "market_price",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IvResult"
                }
              }
            }
          },
          "404": {
            "description": "Unknown underlying"
          },
          "422": {
            "description": "Market price is below intrinsic value"
          }
        }
      }
    },
...
    "/api/v1/price": {
      "get": {
        "tags": [
          "crate"
        ],
        "operationId": "price_option",
        "parameters": [
          {
            "name": "underlying",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "strike",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "expiry_days",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "option_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BSResult"
                }
              }
            }
          },
          "404": {
            "description": "Unknown underlying"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BSResult": {
        "type": "object",
        "required": [
          "premium",
          "delta",
          "gamma",
          "theta",
          "vega",
          "rho",
          "d1",
          "d2",
          "intrinsic",
          "time_value",
          "iv"
        ],
        "properties": {
          "d1": {
            "type": "number",
            "format": "double"
          },
          "d2": {
            "type": "number",
            "format": "double"
          },
          "delta": {
            "type": "number",
            "format": "double"
          },
          "gamma": {
            "type": "number",
            "format": "double"
          },
          "intrinsic": {
            "type": "number",
            "format": "double"
          },
          "iv": {
            "type": "number",
            "format": "double"
          },
          "premium": {
            "type": "number",
            "format": "double"
          },
          "rho": {
            "type": "number",
            "format": "double"
          },
          "theta": {
            "type": "number",
            "format": "double"
          },
          "time_value": {
            "type": "number",
            "format": "double"
          },
          "vega": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "IvResult": {
        "type": "object",
        "required": [
          "implied_vol"
        ],
        "properties": {
          "implied_vol": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "OptionChainEntry": {
        "type": "object",
        "required": [
          "strike",
          "expiry_days",
          "call",
          "put",
          "is_itm_call",
          "is_itm_put"
        ],
        "properties": {
          "call": {
            "$ref": "#/components/schemas/BSResult"
          },
          "expiry_days": {
            "type": "number",
            "format": "double"
          },
          "is_itm_call": {
            "type": "boolean"
          },
          "is_itm_put": {
            "type": "boolean"
          },
          "put": {
            "$ref": "#/components/schemas/BSResult"
          },
          "strike": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque session token"
      }
    }
  }
}